fix(garage): Preserve existing thumbnails on image save failure
- Updated AddEditBikeViewModel and ComponentDetailViewModel to retain existing thumbnail URIs if new image saves fail (return null).
- Enhanced logic to ensure data integrity by preventing loss of thumbnail information during image upload processes.
- Added comments for clarity on the handling of thumbnail updates in case of save failures.

+++ app/src/test/java/com/you/bikecompanion/data/image/ImageRepositoryTest.kt
@@ -26,7 +26,7 @@ class ImageRepositoryTest {
     fun setUp() {
         baseDir = File.createTempFile("image_repo_test", "").apply { delete(); mkdirs() }
         openUriStream = { _ ->
-            ByteArrayInputStream(byteArrayOf(0x89, 0x50, 0x4E, 0x47)) // minimal PNG header
+            ByteArrayInputStream(byteArrayOf(0x89.toByte(), 0x50.toByte(), 0x4E.toByte(), 0x47.toByte())) // minimal PNG header
         }
         repository = ImageRepository(baseDir, openUriStream)
     }
